+++ CHANGELOG.md
@@ -11,6 +11,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 * Fixed `gauge.setToCurrentTime()` to use seconds instead of milliseconds
   * This conforms to Prometheus [best practices](https://prometheus.io/docs/practices/naming/#base-units)
+* Dropped support for node 4
 
 ### Changed